--- scopes/explorer/command-bar/ui/autocomplete-input/index.ts
@@ -1,1 +1,0 @@
-export * from './autocomplete-input';
