+++ scopes/explorer/ui/command-bar/autocomplete-input/autocomplete-input.tsx
