+++ scopes/explorer/ui/command-bar/autocomplete-input/index.ts
@@ -1,0 +1,1 @@
+export * from './autocomplete-input';
