+++ scopes/explorer/ui/command-bar/command-bar-item/command-bar-item.module.scss
