+++ scopes/explorer/ui/command-bar/command-bar-item/command-bar-item.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import classnames from 'classnames';
 import { KeySequence } from '@teambit/ui-foundation.ui.keycap';
 import { mutedText } from '@teambit/base-ui.text.muted-text';
-import { CommanderSearchResult } from '../../types';
+import { CommanderSearchResult } from '../search-result';
 import styles from './command-bar-item.module.scss';
 
 export type CommandItemProps = {
