+++ scopes/explorer/ui/command-bar/command-bar-item/index.ts
