+++ scopes/explorer/ui/command-bar/command-bar/command-bar.module.scss
