+++ scopes/explorer/ui/command-bar/command-bar/index.ts
