+++ scopes/explorer/ui/command-bar/index.ts
@@ -1,0 +1,4 @@
+export { CommandBar } from './command-bar';
+export type { CommandBarProps } from './command-bar';
+
+export { CommanderSearchResult } from './search-result';
