+++ scopes/explorer/ui/command-bar/search-result.ts
@@ -1,0 +1,11 @@
+export type CommandHandler = (...arg: any[]) => any;
+export type Keybinding = string | string[]; // TODO
+
+export type CommanderSearchResult = {
+  id: string;
+  displayName: string;
+  handler: CommandHandler;
+  icon?: string;
+  iconAlt?: string;
+  keybinding?: Keybinding;
+};
